app/pyproject.toml
[tool.poetry]
name = "technical-test"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
argon2-cffi = "^23.1.0"
fastapi = "^0.111.0"
psycopg = "^3.1.19"
psycopg-pool = "^3.2.2"
arq = "^0.26.0"
pytest = "^8.2.2"
asgi-correlation-id = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
